+++ Blasphemous.LostDreams/Timing/TimeHandler.cs
@@ -2,11 +2,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using UnityEngine;
 
-namespace Blasphemous.LostDreams;
+namespace Blasphemous.LostDreams.Timing;
 
-internal class TimeHandler
+internal partial class TimeHandler
 {
     private readonly Dictionary<string, ITimeable> _timers = new();
 
@@ -63,51 +62,5 @@ internal class TimeHandler
     public void Reset()
     {
         RemoveTimeables(x => !x.IsPermanent);
-    }
-
-    class Countdown(float length, Action onFinish) : ITimeable
-    {
-        private readonly Action _onFinish = onFinish;
-        private readonly float _finishTime = Time.time + length;
-
-        public bool ShouldBeRemoved { get; private set; } = false;
-        public bool IsPermanent => false;
-
-        public void OnUpdate()
-        {
-            if (Time.time < _finishTime)
-                return;
-
-            ShouldBeRemoved = true;
-            _onFinish();
-        }
-    }
-
-    class Ticker(float length, bool permanent, Action onTick) : ITimeable
-    {
-        private readonly Action _onTick = onTick;
-        private readonly float _length = length;
-
-        private float _nextActivation = Time.time + length;
-
-        public bool ShouldBeRemoved => false;
-        public bool IsPermanent => permanent;
-
-        public void OnUpdate()
-        {
-            if (Time.time < _nextActivation)
-                return;
-
-            _nextActivation = Time.time + _length;
-            _onTick();
-        }
-    }
-
-    interface ITimeable
-    {
-        bool ShouldBeRemoved { get; }
-        bool IsPermanent { get; }
-
-        void OnUpdate();
     }
 }
